timeseries-ingestion-service/requirements.txt
Flask
uuid
requests
werkzeug
pandas
pyyaml
pypugjs
websocket-client
